backend/requirements.txt
Flask
Flask_Cors
Flask_RESTful
python-decouple
python-dotenv
